refactor customer screen & delete unnecessary line
