+++ .gitignore
@@ -1,0 +1,1 @@
+get_x/.vscode/settings.json
